Added casts around spots where NONE was being assigned to unsigned values
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1968 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/ModelView/Dim3_Loader.cpp
@@ -329,7 +329,7 @@ bool LoadModel_Dim3(FileSpecifier& Spec, Model3D& Model, int WhichPass)
 				if (strncmp(Tag,BoneOwnTags[ibsx].Tag0,BoneTagSize)==0)
 				break;
 			}
-			VS.Bone0 = ibsx < NumBones ? BoneIndices[ibsx] : NONE;
+			VS.Bone0 = ibsx < NumBones ? BoneIndices[ibsx] : static_cast<GLshort>(NONE);
 			
 			Tag = VertexBoneTags[iv].Tag1;
 			for (ibsx=0; ibsx<NumBones; ibsx++)
@@ -337,7 +337,7 @@ bool LoadModel_Dim3(FileSpecifier& Spec, Model3D& Model, int WhichPass)
 				if (strncmp(Tag,BoneOwnTags[ibsx].Tag0,BoneTagSize)==0)
 				break;
 			}
-			VS.Bone1 = ibsx < NumBones ? BoneIndices[ibsx] : NONE;
+			VS.Bone1 = ibsx < NumBones ? BoneIndices[ibsx] : static_cast<GLshort>(NONE);
 		}
 	}
 		
@@ -996,7 +996,7 @@ bool XML_SeqFrameParser::HandleAttribute(const char *Tag, const char *Value)
 		{
 			if (strncmp(Value,FrameTags[ifr].Tag,BoneTagSize) == 0) break;
 		}
-		if (ifr >= NumFrames) ifr = NONE;
+		if (ifr >= NumFrames) ifr = static_cast<size_t>(NONE);
 		Data.Frame = (GLshort)ifr;
 		return true;
 	}
